--- src/components/site-film-section.js
@@ -1,7 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-
-export class FilmSection extends AbstractComponent {
-  getTemplate() {
-    return `<section class="films"></section>`;
-  }
-}
