--- src/components/site-films-commented.js
@@ -1,10 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-
-export class CommentedFilms extends AbstractComponent {
-  getTemplate() {
-    return `
-      <section class="films-list--extra">
-        <h2 class="films-list__title">Most commented</h2>
-      </section>`;
-  }
-}
