--- src/components/site-films-container.js
@@ -1,7 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-
-export class FilmsListContainer extends AbstractComponent {
-  getTemplate() {
-    return `<div class="films-list__container"></div>`;
-  }
-}
