--- src/components/site-films-list-title.js
@@ -1,7 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-
-export class FilmsListTitle extends AbstractComponent {
-  getTemplate() {
-    return `<h2 class="films-list__title"></h2>`;
-  }
-}
