--- src/components/site-films-list.js
@@ -1,7 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-
-export class FilmsList extends AbstractComponent {
-  getTemplate() {
-    return `<section class="films-list"></section>`;
-  }
-}
