--- src/components/site-show-more.js
@@ -1,7 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-
-export class ShowMoreButton extends AbstractComponent {
-  getTemplate() {
-    return `<button class="films-list__show-more">Show more</button>`;
-  }
-}
