--- src/components/site-statistics.js
@@ -1,80 +1,0 @@
-import {AbstractComponent} from "./site-abstract-component";
-import {utils} from "./site-utils";
-
-const mainElement = document.querySelector(`.main`);
-
-export class Statistics extends AbstractComponent {
-  getTemplate() {
-    return `
-    <section class="statistic">
-      <p class="statistic__rank">
-        Your rank 
-        <img class="statistic__img" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35"> 
-        <span class="statistic__rank-label">Sci-Fighter</span>
-      </p>
-    
-      <form action="https://echo.htmlacademy.ru/" method="get" class="statistic__filters">
-        <p class="statistic__filters-description">Show stats:</p>
-    
-        <input type="radio" class="statistic__filters-input visually-hidden" name="statistic-filter" id="statistic-all-time" value="all-time" checked>
-        <label for="statistic-all-time" class="statistic__filters-label">All time</label>
-    
-        <input type="radio" class="statistic__filters-input visually-hidden" name="statistic-filter" id="statistic-today" value="today">
-        <label for="statistic-today" class="statistic__filters-label">Today</label>
-    
-        <input type="radio" class="statistic__filters-input visually-hidden" name="statistic-filter" id="statistic-week" value="week">
-        <label for="statistic-week" class="statistic__filters-label">Week</label>
-    
-        <input type="radio" class="statistic__filters-input visually-hidden" name="statistic-filter" id="statistic-month" value="month">
-        <label for="statistic-month" class="statistic__filters-label">Month</label>
-    
-        <input type="radio" class="statistic__filters-input visually-hidden" name="statistic-filter" id="statistic-year" value="year">
-        <label for="statistic-year" class="statistic__filters-label">Year</label>
-      </form>
-    
-      <ul class="statistic__text-list">
-        <li class="statistic__text-item">
-          <h4 class="statistic__item-title">You watched</h4>
-          <p class="statistic__item-text">22 <span class="statistic__item-description">movies</span></p>
-        </li>
-        <li class="statistic__text-item">
-          <h4 class="statistic__item-title">Total duration</h4>
-          <p class="statistic__item-text">130 <span class="statistic__item-description">h</span> 22 <span class="statistic__item-description">m</span></p>
-        </li>
-        <li class="statistic__text-item">
-          <h4 class="statistic__item-title">Top genre</h4>
-          <p class="statistic__item-text">Sci-Fi</p>
-        </li>
-      </ul>
-    
-      <div class="statistic__chart-wrap">
-        <canvas class="statistic__chart" width="1000"></canvas>
-      </div>
-    </section>`;
-  }
-
-  init() {
-    const statsItem = document.querySelector(`.main-navigation__item--additional`);
-    const navItems = document.querySelectorAll(`.main-navigation__item`);
-
-    statsItem.addEventListener(`click`, (evt) => {
-      evt.preventDefault();
-
-      statsItem.classList.add(`main-navigation__item--active`);
-      document.querySelector(`.films`).classList.add(`visually-hidden`);
-      document.querySelector(`.sort`).classList.add(`visually-hidden`);
-      utils.render(mainElement, AbstractComponent.makeElement(this.getTemplate()), utils.Position.BEFOREEND);
-    });
-
-    navItems.forEach((item) => {
-      item.addEventListener(`click`, (evt) => {
-        evt.preventDefault();
-
-        if (!item.classList.contains(`main-navigation__item--additional`)) {
-          document.querySelector(`.films`).classList.remove(`visually-hidden`);
-          document.querySelector(`.sort`).classList.remove(`visually-hidden`);
-        }
-      });
-    });
-  }
-}
